+++ src/idtrackerai/base/crossings_detection/model_area.py
@@ -3,7 +3,7 @@ import logging
 import numpy as np
 
 from idtrackerai import ListOfBlobs
-from idtrackerai.utils import IdtrackeraiError, conf, track
+from idtrackerai.utils import conf, track
 
 
 class ModelArea:
@@ -40,31 +40,24 @@ class ModelArea:
         logging.info(
             "Initializing ModelArea for individual/crossing blob initial classification"
         )
+        areas = []
         if number_of_animals > 0:
-            areas = []
             for blobs_in_frame in list_of_blobs.blobs_in_video:
                 if len(blobs_in_frame) == number_of_animals:
-                    for blob in blobs_in_frame:
-                        areas.append(blob.area)
-        else:
-            areas = [b.area for b in list_of_blobs.all_blobs]
-        areas = np.asarray(areas)
+                    areas += (blob.area for blob in blobs_in_frame)
 
-        n_blobs = len(areas)
-        if n_blobs == 0:
-            raise IdtrackeraiError(
-                "There is not part in the video where the "
-                f"{number_of_animals} animals are visible. "
-                "Try a different segmentation or check the "
-                "number of animals in the video."
-            )
+        if not areas:
+            # either n_animals is 0 or there are no global fragments
+            areas = [b.area for b in list_of_blobs.all_blobs]
+
+        areas = np.asarray(areas)
         self.median = np.median(areas)
         self.mean = areas.mean()
         self.std = areas.std()
         self.std_tolerance = conf.MODEL_AREA_SD_TOLERANCE
         self.tolerance = self.std_tolerance * self.std
         logging.info(
-            f"Model area computed with {n_blobs} blobs. "
+            f"Model area computed with {len(areas)} blobs. "
             f"Mean area = {self.mean:.1f}, median = {self.median:.1f}, "
             f"and std = {self.std:.1f} (in pixels)"
         )
@@ -77,15 +70,16 @@ def compute_body_length(list_of_blobs: ListOfBlobs, number_of_animals: int) -> f
     """computes the median of the the diagonal of the bounding box."""
     # areas are collected throughout the entire video in the cores of
     # the global fragments
+    body_lengths = []
     if number_of_animals > 0:
-        body_lengths = []
         for blobs_in_frame in track(
             list_of_blobs.blobs_in_video, "Computing body lengths"
         ):
             if len(blobs_in_frame) == number_of_animals:
-                for blob in blobs_in_frame:
-                    body_lengths.append(blob.estimated_body_length)
-    else:
+                body_lengths += (blob.estimated_body_length for blob in blobs_in_frame)
+
+    if not body_lengths:
+        # either n_animals is 0 or there are no global fragments
         body_lengths = [
             b.estimated_body_length
             for b in track(list_of_blobs.all_blobs, "Computing body lengths")