+++ docs/source/user_guide/changelog.rst
@@ -14,6 +14,7 @@ Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 - Refactored and optimized :ref:`idmatcher.ai`.
 - Removed Protocol 3
 - Removed parameters ``THRESHOLD_ACCEPTABLE_ACCUMULATION``, ``MAXIMUM_NUMBER_OF_PARACHUTE_ACCUMULATIONS``, ``MAX_RATIO_OF_PRETRAINED_IMAGES``, and ``protocol3_action``.
+- Tracking is possible without any Global Fragment
 
 5.2.12
 ======